PropProvider Context + Portal Fixes (#454)
This update fixes the transfer for `PropProvider`'s context to content rendered
with `Portal`.

During a React 16 update, we discovered that `PropProvider` values defined
outside of the portal'ed content were not being transferred. This is
due to how `ReactDOM.unstable_renderSubtreeIntoContainer` renders the content
vs. React 16's new context system.

The solution was to create a new HOC (called `contextConnect`), which provides
a component directly with the values from the closest `PropProvider`. Enhancing
`Portal` with this new HOC, I'm able to pass the values through the children
elements. Things work as expected!