Modal + Message.Media Modal updates
This update adjusts the UI for `Modal` to be consistent with the latest
HSDS Design requirements (specifically the `box-shadow` styles).

The `Modal` that renders within `Message.Media` has been adjusted with
better animation defaults to match the experience we have in Beacon Embed.

For added flexibility, these props have been exposed to the component for
refined adjustments.

Lastly, the Story for the `Message.Media` component has been re-written
to include Knobs for a better design/QA experience.