DropdownV2: Resolve item render updates
This update resolves the issue for `DropdownV2` where new items being passed
may (or may not) re-render correctly.

The issue was `Dropdown.Container`, which handles the diff'ing + state
updating process from external props.

The `nextState` that was building was being clobbered over as it trickles
down the conditional logic tree.

---

Unrelated, I fixed the `AnimatedGroup` test so that it stops spitting out
errors during the test run.