Input: Add embedded Action
This update adds the ability for `Input` to render an embedded `Action` within
the Component UI.

Actions (typically [Buttons](../Button)) can be embedded within the `Input` using the `action` prop:

```jsx
<Input
  action={
    <Button version={2} size="sm">
      Submit
    </Button>
  }
/>
```

The `Button` should be one step smaller compared to the `Input`. For example, if the `Input` has a size of `md` (default), then the `Button` should be `sm`. If the `Input` is `xs`, then the `Button` should be `xs`.

Note: Be sure the version for `Button` is at least version 2.