Dropdown: Fix ComboBox positioning with shouldDropDirectionUpdate
This update fixes the up/down menu position of `ComboBox` based dropdowns
that happens when the list size changes (during filtering).

The solution was to introduce a new prop called `shouldDropDirectionUpdate`.
By default, in `Dropdown`, it resolves to `true`. However, in `ComboBox`,
we'll resolve it to `false`.

The user can adjust this, if required.