Icon: Remove/replace 20x20 icons with proper 24x24 svgs
This update removes, replaces, and aliases icons within the `<Icon>` component
to use the latest 24x24 canvas based SVGs from HSDS Core UI.

The **only** icons left that are 20x20 are the ones used to render the caret
UI next to icons when `withCaret` is specified.

The new mini carets are available as `chevron-*` svgs. However, I've decided
to keep the 20x20 caret SVGs for now as we integrate and get a sense of
where/how icon + caret are being used.