Dropdown: Fix top/left position calculation with window.scroll values
This update fixes the issue with Dropdown (V2) where it did not factor in
`window.scrollY` and `window.scrollX` with it's `top` and `left`
calculations.

The fix involved creating a special util function to generated the computed
`top` and `left` values, which are used within the `Dropdown.MenuContainer`
position methods.