Avatar: Render fix for Safari (Mac/iOS)
This update fixes the Avatar image crop rendering for Safari on iOS and Mac OS.
The solution involved migrating the transition styles to a parent DOM node,
in this case, another wrapping `div`, instead of applying it directly on
the `div.image`.

To simplify the (pretty complex) Avatar component, `Avatar.Image` and `Avatar.Crop`
sub-components were created to contain the rendering logic for those pieces
of UI.

This has been tested on both Mac OS and iOS (iPhone 8).