Data Cy: Fix namespace
This update fixes the autogenerated `data-cy` namespace for HSDS: React
components. The solution was to adjust how Fancy compiled the namespace to
make it more compatible with HSDS: React integration.

This required an update to Fancy to the latest v2.4.1

React Utils and Wedux were also updated to their latest versions