Pop (Tooltip/Popover): Fix closeOnBodyClick
This update fixes the closeOnBodyClick mechanic of `Pop`, which powers
`Tooltip` and `Popover`.

The issue was the click handler was never actually called. The solution
was to add `EventListener` to scope on `document.body` to handle the
interaction.

A nice enhancment was added to `Pop` to allow for the user to determine
if the `Pop` component should open/close before it happens with a
`Function` that returns a `boolean`. This enables finer grain control
for situations.

Cyan tests have been added to `Popover` to test these interactions and
integrations.