ActionSelect: Fix resize and refocus interactions
This update fixes the resizing and refocus interactions when an item
is selected from the Dropdown.

A special focus util has been added to ensure the scroll position doesn't
jump on focus.

The resizing mechanism in ActionSelect has been optmized to only fire during
a new selection, rather than on children change.