ActionSelect: Improve UX by coordinating focus + scrolling (#630)
This update improves the UX for the ActionSelect component by scrolling the
content into view when something is selected. The scrolling animation timing
is also synchronized with the height animations, providing the user with a
super smooth experience.

We decided to pursue this enhancement after implementing this component into
the app. When it's positioned with form elements, we found that it was
difficult to (visually) focus on the ActionSelect fields. It felt
cumbersome to have to manually scroll into position.

The auto scroll into view mechanism resolves this issue!