Adjust min-width
If the min-width is 20px, on Firefox there is weird behavior when the
width hits 20px where the text goes from "l...@gmail.com" to
"lon@gmail.com". If you inspect the element, the entire text is still
in the dom, but for some reason only the first three chars are shown
before the splitter. Bumping up the min-width to 21px solves this
problem in Firefox, and the behavior can't be seen in Safari or Chrome.