Add setTimeout to componentDidMount:
When using this component within more complex apps than the
Storybook environment, tooltips were not showing up on the
truncated emails on initial page load because of an issue
where the calculation for isTruncated was being run at a point
where the widths of the contentNode and node were 0. This
caused "isTruncated" in state to be false, which prevented the
tooltip from appearing.

Adding the setTimeout allows this calculation to occur once the
content exists on the page.