Tweak logic for creating boolean 'shouldDiscardEmpty'
Including the check for "removedEmptyFields.length > 0" meant that
when there was a single email visible and the user removed it via
backspace and clicking outside of the input, we would never enter
the code block that handled discarding empty fields. This is b/c
the removedEmptyFields variable ends up being a list of the fields
that remain after an empty field is removed, and in the case of
only having one email in the list, there are no emails left once
you remove it, so 'shouldDiscardEmpty' is always going to be false.