Prevent _id property from being stored in state on delete of final field
if the last visible field has been removed, remove its _id property before
updating the internal state for 'fieldValue' so it accurately represents an
empty field. If _id exists on an otherwise empty field property, it will
force a PUT api call instead of a POST if the user tries to create a new item.