Update data sent to DELETE request so it contains deleted field
Previously, we were filtering the this.state.fieldValue array and
returning the field that did not have a value. However, this filter
function always returned an empty array, because we had previously
updated state so that "fieldValue" only contained the fields with
values present. Instead, we need to pass the "changedField" value
to "onCommit" so that the delete request can be made.