HSAPP-2429 Editable field with multi-values bug (#858)
* Tweak logic for creating boolean 'shouldDiscardEmpty'

Including the check for "removedEmptyFields.length > 0" meant that
when there was a single email visible and the user removed it via
backspace and clicking outside of the input, we would never enter
the code block that handled discarding empty fields. This is b/c
the removedEmptyFields variable ends up being a list of the fields
that remain after an empty field is removed, and in the case of
only having one email in the list, there are no emails left once
you remove it, so 'shouldDiscardEmpty' is always going to be false.

* Prevent _id property from being stored in state on delete of final field

if the last visible field has been removed, remove its _id property before
updating the internal state for 'fieldValue' so it accurately represents an
empty field. If _id exists on an otherwise empty field property, it will
force a PUT api call instead of a POST if the user tries to create a new item.

* Update data sent to DELETE request so it contains deleted field

Previously, we were filtering the this.state.fieldValue array and
returning the field that did not have a value. However, this filter
function always returned an empty array, because we had previously
updated state so that "fieldValue" only contained the fields with
values present. Instead, we need to pass the "changedField" value
to "onCommit" so that the delete request can be made.